\name{aml}
\alias{aml}
\alias{leukemia}
\title{Acute Myelogenous Leukemia survival data}
\usage{data(aml)
data(leukemia)}
\format{
    \tabular{ll}{
	time:\tab survival or censoring time\cr
	status:\tab censoring status\cr
	x: \tab maintenance chemotherapy given? (factor)\cr
     }
 }
\source{Miller "Survival Analysis"}
\keyword{datasets}
