\name{ovarian}
\alias{ovarian}
\title{Ovarian cancer survival data}
\usage{data(ovarian)}
\format{
    \tabular{ll}{
	futime:\tab survival or censoring time\cr
	fustat:\tab censoring status\cr
	age: \tab in years\cr
	resid.ds:\tab residual disease present (1=no,2=yes)\cr
	rx:\tab treatment group\cr
	ecog.ps:\tab ECOG performance status (0=healthy, 5=dead)\cr
     }
 }
\source{Miller "Survival Analysis"}
\keyword{datasets}
